ci: Make python tests runnable in CI. #1460
