+++ py/Makefile
@@ -17,7 +17,7 @@ setup: ## Install dependencies
 	git clone --depth 1 --branch $(WAVE_ML_VERSION) https://github.com/h2oai/wave-ml.git h2o_wave_ml
 	python3 -m venv venv
 	./venv/bin/python -m pip install --upgrade pip
-	./venv/bin/python -m pip install setuptools wheel httpx uvicorn starlette pdoc3 pytest flake8
+	./venv/bin/python -m pip install setuptools wheel httpx uvicorn starlette pdoc3 flake8
 	./venv/bin/python -m pip install h2o_wave_ml/
 	# TODO examples pip install is wasteful for CI
 	./venv/bin/python -m pip install -r examples/requirements.txt
@@ -33,6 +33,9 @@ docs: ## Build API docs
 	mv build/docs/h2o_wave/* ../website/docs/api/
 	mv build/docs/h2o_wave_ml/* ../website/docs/api/h2o_wave_ml/
 	./venv/bin/python sync_examples.py
+
+test:
+	./venv/bin/python -m tests
 
 purge: ## Purge previous build
 	rm -rf build dist h2o_wave.egg-info
