+++ py/tests/__main__.py
@@ -1,0 +1,7 @@
+import unittest
+
+from .test_expando import *
+from .test_python_server import *
+
+if __name__ == '__main__':
+    unittest.main()
